# Backend/go-pos-backend-implementation-ready/docs/openapi.json
{
  "components": {
    "schemas": {
      "Envelope": {
        "properties": {
          "data": {
            "nullable": true,
            "type": "object"
          },
          "error": {
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "request_id": {
                "type": "string"
              }
            },
            "type": "object"
          },
          "meta": {
            "properties": {
              "request_id": {
                "type": "string"
              }
            },
            "type": "object"
          }
        },
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Auto-generated from the Gin route table by cmd/openapi.",
    "title": "pos-api",
    "version": "0.1.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/v1/sync/pull": {
      "get": {
        "tags": [
          "sync"
        ],
        "operationId": "GetV1SyncPull",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            },
            "description": "data/error envelope; see docs/07_API_CONTRACT.md"
          }
        }
      }
    },
    "/v1/sync/push": {
      "post": {
        "tags": [
          "sync"
        ],
        "operationId": "PostV1SyncPush",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            },
            "description": "data/error envelope; see docs/07_API_CONTRACT.md"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080"
    }
  ]
}